Crete result as a Row
Create a row of a result as a Row class instead of plain js object,
to keep parity with datastax driver.

Extend interface of Row constructor, so it can be created
either from array of strings or array of objects
with property name representing column name
(the second options was present in Datastax driver).
