+++ lib/types/results-wrapper.js
@@ -7,6 +7,7 @@ const Duration = require("./duration");
 const LocalTime = require("./local-time");
 const InetAddress = require("./inet-address");
 const { bigintToLong } = require("../new-utils");
+const Row = require("./row");
 
 /**
  * Checks the type of value wrapper object and gets it from the underlying value
@@ -97,7 +98,7 @@ function getRowsFromResultsWrapper(result) {
 
     for (let i = 0; i < rustRows.length; i++) {
         let cols = rustRows[i].getColumns();
-        let collectedRow = {};
+        let collectedRow = new Row(colNames);
         for (let j = 0; j < cols.length; j++) {
             // By default driver returns row as a map:
             //   column name -> column value
