+++ lib/types/row.js
@@ -1,11 +1,15 @@
 "use strict";
 /** @module types */
+
 /**
  * Represents a result row
- * @param {Array} columns
- * @constructor
  */
 class Row {
+    /**
+     * Creates Row from array of column names. Column names can be provided either as array of strings
+     * or array of objects with property name representing column name. Any other metadata will be ignored
+     * @param {Array<string | {name: string}>} columns
+     */
     constructor(columns) {
         if (!columns) {
             throw new Error("Columns not defined");
@@ -24,7 +28,11 @@ class Row {
     get(columnName) {
         if (typeof columnName === "number") {
             // its an index
-            return this[this.__columns[columnName].name];
+            columnName = this.__columns[columnName];
+            if (typeof columnName.name == "string") {
+                columnName = columnName.name;
+            }
+            return this[columnName];
         }
         return this[columnName];
     }
@@ -63,9 +71,5 @@ class Row {
         }
     }
 }
-
-
-
-
 
 module.exports = Row;
